% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Amount_model.R
\name{Amount_model}
\alias{Amount_model}
\title{The Amount model}
\usage{
Amount_model(occur.param, amount.param, rep, obs.data)
}
\arguments{
\item{occur.param}{MC model parameters}

\item{amount.param}{Amount model parameters}

\item{rep}{Number of desired replicates}

\item{obs.data}{Formatted obs data}
}
\value{
TS of simulated rainfall
}
\description{
The Amount model
}
