% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/analyseC.mi.R
\name{analyseC.mi}
\alias{analyseC.mi}
\title{analyseC.mi}
\usage{
analyseC.mi(values1, date1, values2 = NULL, period = NULL)
}
\arguments{
\item{values1,}{format v1,...,vn}

\item{date1}{}

\item{values2,}{format v1,...,vn}

\item{date2}{}

\item{period_seq}{}
}
\description{
MI between variables
}
\examples{
analyseC.mi()
}
\keyword{analysis}
\keyword{classification}

