% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/analyseC.period.R
\name{analyseC.period}
\alias{analyseC.period}
\title{analyseC.period}
\usage{
analyseC.period(date1, period)
}
\arguments{
\item{date1}{}

\item{period}{}
}
\description{
period
}
\examples{
analyseC.period()
}
\keyword{analysis}
\keyword{classification}

