% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/analyseC.R
\name{analyseC}
\alias{analyseC}
\title{analyseC}
\usage{
analyseC(values, dates, action = c("TEMPORAL", "MI"), period = NULL)
}
\arguments{
\item{values,}{format v1,...,vn}

\item{action}{}

\item{date}{}
}
\description{
classification analysis
}
\examples{
analyseC()
}
\keyword{analysis}
\keyword{classification}

