% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/analyseC.readaction.R
\name{analyseC.readaction}
\alias{analyseC.readaction}
\title{analyseC.readaction}
\usage{
analyseC.readaction(action)
}
\arguments{
\item{action}{}
}
\description{
read actions to be done
}
\examples{
analyseC.readaction()
}
\keyword{analysis}
\keyword{classification}

