% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/analyseC.shift.R
\name{analyseC.shift}
\alias{analyseC.shift}
\title{analyseC.shift}
\usage{
analyseC.shift(value, date, shift)
}
\arguments{
\item{date}{}

\item{shift}{}

\item{values}{format v1,...,vn}
}
\description{
Temporal shift of the complete matrix of timeseries
}
\examples{
analyseC.shift()
}
\keyword{analysis}
\keyword{classification}

