% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/analyseTS.basic.R
\name{analyseTS.basic}
\alias{analyseTS.basic}
\title{analyseTS.basic}
\usage{
analyseTS.basic(value, date, filterTime, SHIFT)
}
\arguments{
\item{date}{}

\item{filterTime}{}

\item{SHIFT}{}

\item{values,}{format v1,...,vn}
}
\description{
Basic analysis of a timeseries
}
\examples{
analyseTS.basic()
}
\keyword{entropy}

