% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/analyseTS.class.R
\name{analyseTS.class}
\alias{analyseTS.class}
\title{analyseTS.class}
\usage{
analyseTS.class(valueTS, dateTS, action = "BASIC")
}
\arguments{
\item{valueTS,}{format v1,...,vn}

\item{dateTS}{}

\item{action}{}
}
\description{
classification analysis
}
\examples{
analyseTS.class()
}
\keyword{entropy}

