% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/analyseTS.readaction.R
\name{analyseTS.readaction}
\alias{analyseTS.readaction}
\title{analyseTS.readaction}
\usage{
analyseTS.readaction(action)
}
\arguments{
\item{action}{}
}
\description{
read actions to be done
}
\examples{
analyseTS.readaction()
}
\keyword{entropy}

