% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/analyseTS.shift.R
\name{analyseTS.shift}
\alias{analyseTS.shift}
\title{analyseTS.shift}
\usage{
analyseTS.shift(value, date, shift)
}
\arguments{
\item{date}{}

\item{shift}{}

\item{valeu}{}
}
\description{
Temporal shift of the complete matrix of timeseries
}
\examples{
analyseTS.shift()
}
\keyword{entropy}

