% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/analyseTS.XY.R
\name{analyseTS.XY}
\alias{analyseTS.XY}
\title{analyseTS.XY}
\usage{
analyseTS.XY(X, Y)
}
\arguments{
\item{X}{}

\item{Y}{}
}
\description{
Analysis X and Y
}
\examples{
analyseTS.XY()
}
\keyword{entropy}

