% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/analyseTS.XYfilterTime.R
\name{analyseTS.XYfilterTime}
\alias{analyseTS.XYfilterTime}
\title{analyseTS.XYfilterTime}
\usage{
analyseTS.XYfilterTime(X, Y, dateX, dateY, filterTime)
}
\arguments{
\item{X}{}

\item{Y}{}

\item{dateX}{}

\item{dateY}{}

\item{filterTime}{}
}
\description{
Analysis X and Y for every month or every year or regular one
}
\examples{
analyseTS.XYfilterTime()
}
\keyword{entropy}

