% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/classification.R
\name{classification}
\alias{classification}
\title{classification}
\usage{
classification(geoW, method, TEMPORAL = TRUE, nbclass = c(3, 4), omit,
  FIGURE)
}
\arguments{
\item{geoW}{format geoWeather}

\item{method}{so far som only}

\item{TEMPORAL}{default TRUE}

\item{nbclass}{default c(3,4)}
}
\description{
spatial or temporal classification of a space-time field
}
\examples{
classification()
}
\keyword{classifcation}

