% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/entropy.counting.R
\name{entropy.counting}
\alias{entropy.counting}
\title{entropy.counting}
\usage{
entropy.counting(X, Y)
}
\arguments{
\item{X}{}

\item{Y}{}
}
\description{
counting X and Y
}
\examples{
entropy.counting()
}
\keyword{entropy}

