% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/entropy.d.R
\name{entropy.d}
\alias{entropy.d}
\title{entropy.d}
\usage{
entropy.d(X, Y)
}
\arguments{
\item{X}{}

\item{Y}{}
}
\description{
d(X,Y): Variation of information, d(X,Y)=H(X,Y)-I(X,Y)=H(X)+H(Y)-2I(X,Y)
}
\examples{
entropy.d()
}
\keyword{entropy}

