% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/entropy.DU.R
\name{entropy.DU}
\alias{entropy.DU}
\title{entropy.DU}
\usage{
entropy.DU(X, Y)
}
\arguments{
\item{X}{}

\item{Y}{}
}
\description{
D(X,Y): Universal metric, Jaccard distance, D(X,Y)=d(X,Y)/H(X,Y)=1-I(X,Y)/H(X,Y)
}
\examples{
entropy.DU()
}
\keyword{entropy}

