% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/entropy.hknow.R
\name{entropy.hknow}
\alias{entropy.hknow}
\title{entropy.hknow}
\usage{
entropy.hknow(X, Y)
}
\arguments{
\item{X}{}

\item{Y}{}
}
\description{
H(X/Y)
}
\examples{
entropy.hknow()
}
\keyword{entropy}

