% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/entropy.mi.R
\name{entropy.mi}
\alias{entropy.mi}
\title{entropy.mi}
\usage{
entropy.mi(X, Y)
}
\arguments{
\item{X}{}

\item{Y}{}
}
\description{
I(X,Y): Mutual information, I/X,Y)=H(X,Y)-H(X/Y)-H(Y/X)
}
\examples{
entropy.mi()
}
\keyword{entropy}

