% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/entropy.Pij.R
\name{entropy.Pij}
\alias{entropy.Pij}
\title{entropy.Pij}
\usage{
entropy.Pij(X, Y)
}
\arguments{
\item{X}{}

\item{Y}{}
}
\description{
Probability P(i,j)
}
\examples{
entropy.Pij()
}
\keyword{entropy}

