% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/entropy.Xsumup.R
\name{entropy.Xsumup}
\alias{entropy.Xsumup}
\title{entropy.Xsumup}
\usage{
entropy.Xsumup(X)
}
\arguments{
\item{X}{}
}
\description{
sum-up of X
}
\examples{
entropy.Xsumup()
}
\keyword{entropy}

