% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/entropy.XY.R
\name{entropy.XY}
\alias{entropy.XY}
\title{entropy.XY}
\usage{
entropy.XY(X, Y)
}
\arguments{
\item{X}{}

\item{Y}{}
}
\description{
entropy between X and Y
}
\examples{
entropy.XY()
}
\keyword{entropy}

