% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitAmountModel.R
\name{fitAmountModel}
\alias{fitAmountModel}
\title{Calibrate the amount model: a gamma or a exponential distribution}
\usage{
fitAmountModel(obs.data, mod)
}
\arguments{
\item{obs.data}{observed daily TS of rainfall}

\item{mod}{User specify a gamma or a exponential distribution}
}
\value{
Amount model parameters
}
\description{
Calibrate the amount model: a gamma or a exponential distribution
}
