% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitMarkovModel.R
\name{fitMCModel}
\alias{fitMCModel}
\title{Calibrate the 2-state Markov Cahin model parameters PWW and PDW}
\usage{
fitMCModel(obs.data)
}
\arguments{
\item{obs.data}{observed daily TS of rainfall}
}
\value{
MC model parameters PWW and PDW
}
\description{
Calibrate the 2-state Markov Cahin model parameters PWW and PDW
}
