% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormatTimeSeries.R
\name{format_TimeSeries}
\alias{format_TimeSeries}
\title{Formatting a rainfall time series in to day, month, year}
\usage{
format_TimeSeries(x)
}
\arguments{
\item{x}{the Time Series}
}
\value{
A mutated Time Series with new columns for day, month and year
}
\description{
Formatting a rainfall time series in to day, month, year
}
