% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSimRain.R
\name{getSimRain}
\alias{getSimRain}
\title{Get simulated rainfall}
\usage{
getSimRain(obs.data, rep = 100, mod = "gama")
}
\arguments{
\item{rep}{Number of desired replicates}

\item{mod}{Specified "gama" or "expo" distribution}

\item{occur.param}{MC model parameters}

\item{amount.param}{Amount model parameters}
}
\value{
TS of simulated rainfall
}
\description{
Get simulated rainfall
}
