% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/meteorology.e.R
\name{meteorology.e}
\alias{meteorology.e}
\title{meteorology.e}
\usage{
meteorology.e(Td)
}
\arguments{
\item{Td}{}
}
\description{
the non-staturated and saturated vapour pressure; expression based on Bolton(1980)
REM: Other more complex formulae for saturation vapour pressure on liquid water and on ice can be obtained through polinomical adjustments (Flatan et al. 1992)
}
\examples{
meteorology.e()
}
\keyword{meteorology}

