% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/meteorology.KI.R
\name{meteorology.KI}
\alias{meteorology.KI}
\title{meteorology.KI}
\usage{
meteorology.KI(T, Td, p)
}
\arguments{
\item{T}{}

\item{Td}{}

\item{p}{}
}
\description{
KI: K-index: (850 hPa temperature - 500 hPa temperature) + 850 hPa dew point - 700 hPa dew point depression
}
\examples{
meteorology.KI()
}
\keyword{meteorology}

