% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/meteorology.lcl.R
\name{meteorology.lcl}
\alias{meteorology.lcl}
\title{meteorology.lcl}
\usage{
meteorology.lcl(T, mr, p)
}
\arguments{
\item{T}{}

\item{mr}{}

\item{p}{}
}
\description{
lcl: Lifted Condensation Level, pressure and temperature
Because shallow convection mixes the boundry layer before afternoon convection is initiated,
it makes sense to graphically estimate mean values of mixing ratio and potential temperature and
to use them to determine the LCL, LFC, and EL.
}
\examples{
meteorology.lcl()
}
\keyword{meteorology}

