% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/meteorology.lfcel.R
\name{meteorology.lfcel}
\alias{meteorology.lfcel}
\title{meteorology.lfcel}
\usage{
meteorology.lfcel(Twbp, plcl, T, p)
}
\arguments{
\item{Twbp}{}

\item{plcl}{}

\item{T}{}

\item{p}{}
}
\description{
lfc: Level of Free Convection
el: Equilibrium Level
Because shallow convection mixes the boundry layer before afternoon convection is initiated,
it makes sense to graphically estimate mean values of mixing ratio and potential temperature and
to use them to determine the LCL, LFC, and EL.
}
\examples{
meteorology.lfcel()
}
\keyword{meteorology}

