% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/meteorology.LI.R
\name{meteorology.LI}
\alias{meteorology.LI}
\title{meteorology.LI}
\usage{
meteorology.LI(T, Twbp, p)
}
\arguments{
\item{T}{}

\item{Twbp}{}

\item{p}{}
}
\description{
LI: Lifted Index: T500-Tmoistadiab500
}
\examples{
meteorology.LI()
}
\keyword{meteorology}

