% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/meteorology.pw.R
\name{meteorology.pw}
\alias{meteorology.pw}
\title{meteorology.pw}
\usage{
meteorology.pw(p, Td, T, minp = 400)
}
\arguments{
\item{p}{}

\item{Td}{}

\item{T}{}

\item{minp}{}
}
\description{
precipitable water (in mm) up to a minimum pressure
}
\examples{
meteorology.pw()
}
\keyword{meteorology}

