% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/meteorology.sh.R
\name{meteorology.sh}
\alias{meteorology.sh}
\title{meteorology.sh}
\usage{
meteorology.sh(e, p)
}
\arguments{
\item{e}{}

\item{p}{}
}
\description{
the specific humidity, in function of the water vapor pressure of the saturated air (e) and the pressure of the saturated air (p-e)
}
\examples{
meteorology.sh()
}
\keyword{meteorology}

