% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/meteorology.Tda.R
\name{meteorology.Tda}
\alias{meteorology.Tda}
\title{meteorology.Tda}
\usage{
meteorology.Tda(T, p)
}
\arguments{
\item{T}{}

\item{p}{}
}
\description{
temperature on the dry adiabatic; poisson's equation
}
\examples{
meteorology.Tda()
}
\keyword{meteorology}

