% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/meteorology.Tp.R
\name{meteorology.Tp}
\alias{meteorology.Tp}
\title{meteorology.Tp}
\usage{
meteorology.Tp(T, p)
}
\arguments{
\item{T}{}

\item{p}{}
}
\description{
Potential temperature: temperature that the parcel would acquire if adiabatically brought to a standard reference pressure
}
\examples{
meteorology.Tp()
}
\keyword{meteorology}

