% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/meteorology.Tv.R
\name{meteorology.Tv}
\alias{meteorology.Tv}
\title{meteorology.Tv}
\usage{
meteorology.Tv(T, mr)
}
\arguments{
\item{T}{}

\item{mr}{}
}
\description{
virtual temperature
}
\examples{
meteorology.Tv()
}
\keyword{meteorology}

