% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/meteorology.Twb.R
\name{meteorology.Twb}
\alias{meteorology.Twb}
\title{meteorology.Twb}
\usage{
meteorology.Twb(T, mr, p)
}
\arguments{
\item{T}{}

\item{mr}{}

\item{p}{}
}
\description{
Wet bulb temperature
}
\examples{
meteorology.Twb()
}
\keyword{meteorology}

