% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/meteorology.Twbp.R
\name{meteorology.Twbp}
\alias{meteorology.Twbp}
\title{meteorology.Twbp}
\usage{
meteorology.Twbp(T, p)
}
\arguments{
\item{T}{}

\item{p}{}
}
\description{
Potential Wet bulb temperature
}
\examples{
meteorology.Twbp()
}
\keyword{meteorology}

