% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/meteorology.wobf.R
\name{meteorology.wobf}
\alias{meteorology.wobf}
\title{meteorology.wobf}
\usage{
meteorology.wobf(T)
}
\arguments{
\item{T}{}
}
\description{
wobf
COPYRIGHT REMARK:  Due to a lack of time, this function has been adapted from the R-package "radiosonde"
http://www.image.ucar.edu/Software/RadioSonde/
}
\examples{
meteorology.wobf()
}
\keyword{meteorology}

