% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/meteorology.z.R
\name{meteorology.z}
\alias{meteorology.z}
\title{meteorology.z}
\usage{
meteorology.z(z, T)
}
\arguments{
\item{z}{}

\item{T}{}
}
\description{
elevation
}
\examples{
meteorology.z()
}
\keyword{meteorology}

