% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Occurence_model.R
\name{monthly_occurr_model}
\alias{monthly_occurr_model}
\title{Markov Chain First order model to generate a binary series for monthly rainfall occurrence}
\usage{
monthly_occurr_model(N, PDW, PWW)
}
\arguments{
\item{N}{length of obs monthly rainfall}

\item{PDW}{transitional probability}

\item{PWW}{transitional probability}
}
\value{
The binary series of monthly rainfall occurence
}
\description{
Markov Chain First order model to generate a binary series for monthly rainfall occurrence
}
