% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/score.ContingencyCase.R
\name{score.ContingencyCase}
\alias{score.ContingencyCase}
\title{score.ContingencyCase}
\usage{
score.ContingencyCase(data1, data2, ths1, ths2, caseCT = "HIT")
}
\arguments{
\item{data1}{}

\item{data2}{}

\item{ths1}{threshold#1}

\item{ths2}{threshold#2}

\item{caseCT}{default "HIT" #"FA", "MISSING", "NEG"}
}
\description{
score.ContingencyCase
}
\examples{
score.ContingencyCase()
}
\keyword{score}

