% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/score.CSI_b.R
\name{score.CSI_b}
\alias{score.CSI_b}
\title{score.CSI_b}
\usage{
score.CSI_b(matrix_data, matrix_data2, ths)
}
\arguments{
\item{matrix_data}{observed}

\item{matrix_data2}{forecasted}

\item{ths}{}
}
\description{
Critical Success Index
The CSI, also known as threat score, ranges from 0 to 1 where 0 indicates no skill and 1 is a perfect score. It
measures the fraction of the observed and/or forecasted precipitation that was correctly predicted. Unlike the
POD and the FAR, it takes into account both false alarms and misses, and is therefore a more balanced score
}
\examples{
score.CSI_b()
}
\keyword{score}

