% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/score.FA.R
\name{score.FA}
\alias{score.FA}
\title{score.FA}
\usage{
score.FA(data1, data2, ths1, ths2)
}
\arguments{
\item{data1}{}

\item{data2}{}

\item{ths1}{threshold#1}

\item{ths2}{threshold#2}
}
\description{
False Alarm
}
\examples{
score.FA()
}
\keyword{score}

