% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/score.HIT.R
\name{score.HIT}
\alias{score.HIT}
\title{score.HIT}
\usage{
score.HIT(data1, data2, ths1, ths2)
}
\arguments{
\item{data1}{}

\item{data2}{}

\item{ths1}{threshold#1}

\item{ths2}{threshold#2}
}
\description{
Hit
}
\examples{
score.HIT()
}
\keyword{score}

