% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/score.MISSING.R
\name{score.MISSING}
\alias{score.MISSING}
\title{score.MISSING}
\usage{
score.MISSING(data1, data2, ths1, ths2)
}
\arguments{
\item{data1}{}

\item{data2}{}

\item{ths1}{threshold#1}

\item{ths2}{threshold#2}
}
\description{
Missing
}
\examples{
score.MISSING()
}
\keyword{score}

