% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/score.SEEPS_b.R
\name{score.SEEPS_b}
\alias{score.SEEPS_b}
\title{score.SEEPS_b}
\usage{
score.SEEPS_b(matrix_data, matrix_data2, p1, ths)
}
\arguments{
\item{matrix_data}{observed}

\item{matrix_data2}{forecasted}

\item{ths}{}
}
\description{
Seeps
score SEEPS, cf Rodwell et al 2010; Haiden et al 2012
0:perfect
the higher value, the less skilled
}
\examples{
score.SEEPS_b()
}
\keyword{score}

