% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/skewT.ab.R
\name{skewT.ab}
\alias{skewT.ab}
\title{skewT.ab}
\usage{
skewT.ab(x1, x2, y1, y2)
}
\arguments{
\item{x1}{}

\item{x2}{}

\item{y1}{}

\item{y2}{}
}
\description{
a,b slope and coef
}
\examples{
skewT.ab()
}
\keyword{skewT}

