% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/skewT.line.R
\name{skewT.line}
\alias{skewT.line}
\title{skewT.line}
\usage{
skewT.line(p_min, p_max, T_min, T_max, T, p, T_name)
}
\arguments{
\item{p_min}{}

\item{p_max}{}

\item{T_min}{}

\item{T_max}{}

\item{T}{}

\item{p}{}

\item{T_name}{}
}
\description{
plot a Temperature radiosounding on the skewT graph, choice: Temperature, Wet bulb temperature, Dew point temperature, Potential Wet bulb temperature, Virtual Temperature
}
\examples{
skewT.line()
}
\keyword{skewT}

