% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/skewT.plot.R
\name{skewT.plot}
\alias{skewT.plot}
\title{skewT.plot}
\usage{
skewT.plot(p_min, p_max, T_min, T_max)
}
\arguments{
\item{p_min}{}

\item{p_max}{}

\item{T_min}{}

\item{T_max}{}
}
\description{
plot an empty skewT graph
}
\examples{
skewT.plot()
}
\keyword{skewT}

