% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/skewT.point.R
\name{skewT.point}
\alias{skewT.point}
\title{skewT.point}
\usage{
skewT.point(p_min, p_max, T_min, T_max, T, p)
}
\arguments{
\item{p_min}{}

\item{p_max}{}

\item{T_min}{}

\item{T_max}{}

\item{T}{}

\item{p}{}
}
\description{
plot a Temperature point on the skewT graph, choice: Temperature, Wet bulb temperature, Dew point temperature, Potential Wet bulb temperature, Virtual Temperature
}
\examples{
skewT.point()
}
\keyword{skewT}

