% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/skewT.x.R
\name{skewT.x}
\alias{skewT.x}
\title{skewT.x}
\usage{
skewT.x(a, b, T_min, T, p)
}
\arguments{
\item{a}{}

\item{b}{}

\item{T_min}{}

\item{T}{}

\item{p}{}
}
\description{
plot the x-axis of an empty skewT graph
}
\examples{
skewT.x()
}
\keyword{skewT}

