% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/skewT.y.R
\name{skewT.y}
\alias{skewT.y}
\title{skewT.y}
\usage{
skewT.y(a, b, p)
}
\arguments{
\item{a}{}

\item{b}{}

\item{p}{}
}
\description{
plot the y-axis of an empty skewT graph
}
\examples{
skewT.y()
}
\keyword{skewT}

