% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/thermodynamics.e.R
\name{thermodynamics.e}
\alias{thermodynamics.e}
\title{thermodynamics.e}
\usage{
thermodynamics.e(Td)
}
\arguments{
\item{Td}{}
}
\description{
the non-staturated and saturated vapour pressure; expression based on Bolton(1980)
REM: Other more complex formulae for saturation vapour pressure on liquid water and on ice can be obtained through polinomical adjustments (Flatan et al. 1992)
}
\examples{
thermodynamics.e()
}
\keyword{thermodynamics}

