% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/thermodynamics.es.R
\name{thermodynamics.es}
\alias{thermodynamics.es}
\title{thermodynamics.es}
\usage{
thermodynamics.es(T)
}
\arguments{
\item{T}{}
}
\description{
the staturated and saturated vapour pressure; expression based on Bolton(1980)
REM: Other more complex formulae for saturation vapour pressure on liquid water and on ice can be obtained through polinomical adjustments (Flatan et al. 1992)
}
\examples{
thermodynamics.es()
}
\keyword{thermodynamics}

