% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/thermodynamics.KI.R
\name{thermodynamics.KI}
\alias{thermodynamics.KI}
\title{thermodynamics.KI}
\usage{
thermodynamics.KI(T, Td, p)
}
\arguments{
\item{T}{}

\item{Td}{}

\item{p}{}
}
\description{
KI: K-index: (850 hPa temperature - 500 hPa temperature) + 850 hPa dew point - 700 hPa dew point depression
}
\examples{
thermodynamics.KI()
}
\keyword{thermodynamics}

