% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/thermodynamics.LI.R
\name{thermodynamics.LI}
\alias{thermodynamics.LI}
\title{thermodynamics.LI}
\usage{
thermodynamics.LI(T, Twbp, p)
}
\arguments{
\item{T}{}

\item{Twbp}{}

\item{p}{}
}
\description{
LI: Lifted Index: T500-Tmoistadiab500
}
\examples{
thermodynamics.LI()
}
\keyword{thermodynamics}

