% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/thermodynamics.moistadiab.R
\name{thermodynamics.moistadiab}
\alias{thermodynamics.moistadiab}
\title{thermodynamics.moistadiab}
\usage{
thermodynamics.moistadiab(T, p)
}
\arguments{
\item{T}{}

\item{p}{}
}
\description{
saturated adiabatic
COPYRIGHT REMARK:  Due to a lack of time, this function has been adapted from the R-package "radiosonde"
http://www.image.ucar.edu/Software/RadioSonde/
}
\examples{
thermodynamics.moistadiab()
}
\keyword{thermodynamics}

