% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/thermodynamics.mr.R
\name{thermodynamics.mr}
\alias{thermodynamics.mr}
\title{thermodynamics.mr}
\usage{
thermodynamics.mr(e, p)
}
\arguments{
\item{e}{}

\item{p}{}
}
\description{
the mixing ratio of the mass vapor to the mass of dry air, in function of the water vapor pressure of the saturated air (e) and the pressure of the saturated air (p-e)
}
\examples{
thermodynamics.mr()
}
\keyword{thermodynamics}

