% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/thermodynamics.pw.R
\name{thermodynamics.pw}
\alias{thermodynamics.pw}
\title{thermodynamics.pw}
\usage{
thermodynamics.pw(p, Td, T, minp = 400)
}
\arguments{
\item{p}{}

\item{Td}{}

\item{T}{}

\item{minp}{}
}
\description{
precipitable water (in mm) up to a minimum pressure
}
\examples{
thermodynamics.pw()
}
\keyword{thermodynamics}

