% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/thermodynamics.sh.R
\name{thermodynamics.sh}
\alias{thermodynamics.sh}
\title{thermodynamics.sh}
\usage{
thermodynamics.sh(e, p)
}
\arguments{
\item{e}{}

\item{p}{}
}
\description{
the specific humidity, in function of the water vapor pressure of the saturated air (e) and the pressure of the saturated air (p-e)
}
\examples{
thermodynamics.sh()
}
\keyword{thermodynamics}

