% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/thermodynamics.Td.R
\name{thermodynamics.Td}
\alias{thermodynamics.Td}
\title{thermodynamics.Td}
\usage{
thermodynamics.Td(rh, T)
}
\arguments{
\item{rh}{}

\item{T}{}
}
\description{
Dew Point Temperature
}
\examples{
thermodynamics.Td()
}
\keyword{thermodynamics}

