% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/thermodynamics.Tda.R
\name{thermodynamics.Tda}
\alias{thermodynamics.Tda}
\title{thermodynamics.Tda}
\usage{
thermodynamics.Tda(T, p)
}
\arguments{
\item{T}{}

\item{p}{}
}
\description{
temperature on the dry adiabatic; poisson's equation
}
\examples{
thermodynamics.Tda()
}
\keyword{thermodynamics}

