% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/thermodynamics.Tmrp.R
\name{thermodynamics.Tmrp}
\alias{thermodynamics.Tmrp}
\title{thermodynamics.Tmrp}
\usage{
thermodynamics.Tmrp(mr, p)
}
\arguments{
\item{mr}{}

\item{p}{}
}
\description{
temperature on a given mr and p; table 1 on page 7 of stipanuk (1973)
}
\examples{
thermodynamics.Tmrp()
}
\keyword{thermodynamics}

