% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/thermodynamics.Tp.R
\name{thermodynamics.Tp}
\alias{thermodynamics.Tp}
\title{thermodynamics.Tp}
\usage{
thermodynamics.Tp(T, p)
}
\arguments{
\item{T}{}

\item{p}{}
}
\description{
Potential temperature: temperature that the parcel would acquire if adiabatically brought to a standard reference pressure
}
\examples{
thermodynamics.Tp()
}
\keyword{thermodynamics}

