% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/thermodynamics.Tv.R
\name{thermodynamics.Tv}
\alias{thermodynamics.Tv}
\title{thermodynamics.Tv}
\usage{
thermodynamics.Tv(T, mr)
}
\arguments{
\item{T}{}

\item{mr}{}
}
\description{
virtual temperature
}
\examples{
thermodynamics.Tv()
}
\keyword{thermodynamics}

