% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/thermodynamics.Twb.R
\name{thermodynamics.Twb}
\alias{thermodynamics.Twb}
\title{thermodynamics.Twb}
\usage{
thermodynamics.Twb(T, mr, p)
}
\arguments{
\item{T}{}

\item{mr}{}

\item{p}{}
}
\description{
Wet bulb temperature
}
\examples{
thermodynamics.Twb()
}
\keyword{thermodynamics}

