% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/thermodynamics.Twbp.R
\name{thermodynamics.Twbp}
\alias{thermodynamics.Twbp}
\title{thermodynamics.Twbp}
\usage{
thermodynamics.Twbp(T, p)
}
\arguments{
\item{T}{}

\item{p}{}
}
\description{
Potential Wet bulb temperature
}
\examples{
thermodynamics.Twbp()
}
\keyword{thermodynamics}

