% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/thermodynamics.wobf.R
\name{thermodynamics.wobf}
\alias{thermodynamics.wobf}
\title{thermodynamics.wobf}
\usage{
thermodynamics.wobf(T)
}
\arguments{
\item{T}{}
}
\description{
wobf
COPYRIGHT REMARK:  Due to a lack of time, this function has been adapted from the R-package "radiosonde"
http://www.image.ucar.edu/Software/RadioSonde/
}
\examples{
thermodynamics.wobf()
}
\keyword{thermodynamics}

