% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/thermodynamics.z.R
\name{thermodynamics.z}
\alias{thermodynamics.z}
\title{thermodynamics.z}
\usage{
thermodynamics.z(p, T)
}
\arguments{
\item{p}{}

\item{T}{}
}
\description{
elevation
}
\examples{
thermodynamics.z()
}
\keyword{thermodynamics}

