% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGEN_Package_Description.R
\docType{package}
\name{WGEN}
\alias{WGEN}
\title{WGEN Stochastic Rainfall Model Package}
\value{
A List containing 12 dataframes (respective to 12 months) each dataframe contains N replicates (N: user-specified) of monthly rainfall timeseries
}
\description{
Simulate rainfall timeseries from observed rainfall data. The observed rainfall data is required to be formatted prior the simulation using the built-in format_TimeSeries() function. Example: x <- format_Timeseries(obs.data); simRain <- getSimRain(x)
}
\author{
Thien (JAMES) Nguyen
}
